refactor(ScreenshotContext): use html-to-image for lighter bundle, faster processing
